+++ src/StartUpDG.jl
@@ -6,60 +6,43 @@ Module to aid in setting up reference operators, meshes, and geometric terms
 
 module StartUpDG
 
-using NodesAndModes # for basis functions
 using LinearAlgebra # for diagm, identity matrix I
-using StaticArrays
-using UnPack        # for getting values in RefElemData and MeshData
+using NodesAndModes # for basis functions
+using StaticArrays  # for SMatrix
 using Setfield      # for "modifying" structs to modify node mappings
+using UnPack        # for getting values in RefElemData and MeshData
 
+import NodesAndModes: meshgrid
 import SparseArrays: sparse, droptol!
 import UnPack: @unpack
 
-import NodesAndModes: meshgrid
-
 export Line, Tri, Quad, Hex # element types from NodesAndModes
-
 export @unpack
 export meshgrid
 
-# convenience containers for reference element and physical data types.
-export RefElemData
-export MeshData
-include("DG_types.jl")
-
-# initialization of mesh/reference element data
-export init_reference_elem
-include("ref_elem_functions.jl")
-
-# WARNING: experimental
+# reference element utility functions
 export RefElemData2
-include("ref_elem_data.jl")
+include("RefElemData.jl")
+include("ref_elem_utils.jl")
+
 export MeshData2
-include("mesh_data.jl")
+include("MeshData.jl")
 
 # spatial assembly routines
-export connect_mesh
-export build_node_maps, geometric_factors
-export build_periodic_boundary_maps, build_periodic_boundary_maps!
-export init_DG_mesh
-include("./mesh_functions.jl")
-include("./geometric_mapping_functions.jl")
-include("./node_connectivity_functions.jl")
+export connect_mesh, build_node_maps
+export build_periodic_boundary_maps!
+include("./connectivity_functions.jl")
 
 # uniform meshes + face vertex orderings are included
-export readGmsh2D
-export uniform_mesh
-export face_vertices
-include("mesh_utils.jl")
+export readGmsh2D, uniform_mesh
+include("simple_meshes.jl")
 
-# submodule for explicit time-stepping included for convenience
-export ExplicitTimestepUtils
-module ExplicitTimestepUtils
-import ..@unpack
-export bcopy!, bmult
-export ck45, dp56 # carpenter/kennedy and dormand/prince
-export PIparams, init_PI_controller, compute_adaptive_dt # for embedded RK methods like dp56()
+# ref-to-physical geometric terms
+export geometric_factors
+include("./geometric_mapping_functions.jl")
+
+# simple explicit time-stepping included for conveniencea
+export ck45, dp56, PIparams, compute_adaptive_dt # LSERK 45 + Dormand-Prince 56
 include("explicit_timestep_utils.jl")
-end
 
 end
