+++ test/runtests.jl
@@ -2,21 +2,15 @@ using StartUpDG
 using Test
 using LinearAlgebra
 
-@testset "Timestep utils" begin
+@testset "Timestep and mesh utils" begin
     tol = 5e2*eps()
 
     EToV,VX,VY = readGmsh2D("squareCylinder2D.msh")
     @test size(EToV)==(3031,3)
 
-    using StartUpDG.ExplicitTimestepUtils
-    a = ntuple(x->randn(2,3),3)
-    b = ntuple(x->randn(2,3),3)
-    bcopy!.(a,b)
-    @test a==b
-
     # feed zero rhs to PI controller = max timestep, errEst = 0
     rka,rkE,rkc = dp56()
-    PI = init_PI_controller(5)
+    PI = PIparams(order=5)
     Q = (randn(2,4),randn(2,4))
     rhsQrk = ntuple(x->zero.(Q),length(rkE))
     accept_step, dt_new, errEst =
